+++ include/metalang99/util.h
@@ -65,7 +65,7 @@
 /**
  * Evaluates to nothing.
  */
-#define ML99_empty() ML99_callUneval(ML99_empty, )
+#define ML99_empty(...) ML99_callUneval(ML99_empty, )
 
 /**
  * Evaluates to its arguments.
@@ -108,20 +108,6 @@
  * @endcode
  */
 #define ML99_flip(f) ML99_call(ML99_flip, f)
-
-/**
- * Consumes all its arguments and expands to emptiness.
- *
- * # Examples
- *
- * @code
- * #include <metalang99/util.h>
- *
- * // ML99_empty()
- * ML99_consume(v(1, 2, 3))
- * @endcode
- */
-#define ML99_consume(...) ML99_call(ML99_consume, __VA_ARGS__)
 
 /**
  * Concatenates @p x with @p y as-is, without expanding them.
@@ -198,19 +184,17 @@
 #define ML99_STRINGIFY(...) ML99_STRINGIFY_PRIMITIVE(__VA_ARGS__)
 #define ML99_EMPTY()
 #define ML99_ID(...) __VA_ARGS__
-#define ML99_CONSUME(...)
 
 #ifndef DOXYGEN_IGNORE
 
 #define ML99_catEval_IMPL(x, y)      x##y
 #define ML99_cat_IMPL(x, y)          v(ML99_CAT(x, y))
 #define ML99_stringify_IMPL(...)     v(ML99_STRINGIFY(__VA_ARGS__))
-#define ML99_empty_IMPL()            v(ML99_EMPTY())
+#define ML99_empty_IMPL(...)         v(ML99_EMPTY())
 #define ML99_id_IMPL(...)            v(ML99_ID(__VA_ARGS__))
 #define ML99_const_IMPL(x, _a)       v(x)
 #define ML99_flip_IMPL(f)            ML99_appl_IMPL(ML99_PRIV_flip, f)
 #define ML99_PRIV_flip_IMPL(f, a, b) ML99_appl2_IMPL(f, b, a)
-#define ML99_consume_IMPL(...)       v(ML99_CONSUME(__VA_ARGS__))
 
 #if defined(__GNUC__) && !defined(__clang__)
 #define ML99_PRIV_GCC_PRAGMA(str) _Pragma(str)
@@ -232,7 +216,6 @@
 #define ML99_id_ARITY        1
 #define ML99_const_ARITY     2
 #define ML99_flip_ARITY      1
-#define ML99_consume_ARITY   1
 
 #define ML99_PRIV_flip_ARITY 3
 // }