+++ tests/control.c
@@ -4,11 +4,11 @@
 #include <metalang99/util.h>
 
 // ML99_if {
-ML99_ASSERT_EQ(ML99_if(ML99_true, v(24), v(848)), v(24));
-ML99_ASSERT_EQ(ML99_if(ML99_true, v(1549), v(1678)), v(1549));
+ML99_ASSERT_EQ(ML99_if(ML99_true(), v(24), v(848)), v(24));
+ML99_ASSERT_EQ(ML99_if(ML99_true(), v(1549), v(1678)), v(1549));
 
-ML99_ASSERT_EQ(ML99_if(ML99_false, v(516), v(115)), v(115));
-ML99_ASSERT_EQ(ML99_if(ML99_false, v(10), v(6)), v(6));
+ML99_ASSERT_EQ(ML99_if(ML99_false(), v(516), v(115)), v(115));
+ML99_ASSERT_EQ(ML99_if(ML99_false(), v(10), v(6)), v(6));
 // }
 
 // ML99_IF {
@@ -18,8 +18,8 @@ ML99_ASSERT_EQ(ML99_if(ML99_false, v(10), v(6)), v(6));
 #define X 1, 2, 3
 
 // Test multiple commas in branches.
-CHECK(ML99_IF(ML99_TRUE, X, ~));
-CHECK(ML99_IF(ML99_FALSE, ~, X));
+CHECK(ML99_IF(ML99_TRUE(), X, ~));
+CHECK(ML99_IF(ML99_FALSE(), ~, X));
 
 #undef CHECK
 #undef CHECK_AUX