+++ CHANGELOG.md
@@ -11,11 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - `ML99_QUOTE`
 
+### Removed
+
+ - `ML99_consume`, `ML99_CONSUME`
+
 ### Changed
 
  - Move `ML99_TERMS` from `util.h` to `lang.h`.
  - Return a list of tuples from `ML99_listZip`, accept a list of tuples in `ML99_listUnzip`, return a tuple of lists from `ML99_listPartition`.
  - `ML99_listEval` -> `ML99_LIST_EVAL`, `ML99_listEvalCommaSep` -> `ML99_LIST_EVAL_COMMA_SEP`
+ - Accept variadic arguments in `ML99_nil`, `ML99_empty`, `ML99_true`, `ML99_TRUE`, `ML99_false`, `ML99_FALSE` and ignore them.
 
 ### Fixed